requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy==1.18.5
pandas
scikit-learn==0.22

# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs
google-cloud-storage
mlflow
s3fs

# utilities
six==1.15.0
joblib
memoized-property
termcolor

python-gitlab
